refactor tests + readme
