+++ elasticai/creator/systemTests/brevitas_representation/test_brevitas_representation.py
@@ -9,7 +9,7 @@ from elasticai.creator.brevitas.brevitas_model_comparison import (
 from elasticai.creator.brevitas.brevitas_representation import (
     BrevitasRepresentation,
 )
-from elasticai.creator.systemTests.models_definition import (
+from elasticai.creator.systemTests.brevitas_representation.models_definition import (
     create_qtorch_model,
     create_brevitas_model,
 )
