+++ elasticai/creator/tests/vhdl/test_language_testbench.py
@@ -1,5 +1,5 @@
 from elasticai.creator.vhdl.language_testbench import (
-    TestCasesProcomputedScalarFunction,
+    TestCasesPrecomputedScalarFunction,
     TestCasesLSTMCommonGate,
     TestCasesLSTMCell,
     Procedure,
@@ -168,4 +168,53 @@ class TestCasesLSTMCommonGateTest(TestCase):
         )
 
 
-# TODO: test cases for precomputed scalar function testcases
+class TestCasesPrecomputedScalarFunctionTest(TestCase):
+    def test_TestCasesPrecomputedScalarFunction(self):
+        x_list_for_testing = [1, 2, 3]
+        y_list_for_testing = [-5, "some_string", -3]
+        x_variable_name = "x_name"
+        y_variable_name = "y_name"
+        data_width = 3
+        test_cases_precomputed_scalar_function = TestCasesPrecomputedScalarFunction(
+            x_list_for_testing=x_list_for_testing,
+            y_list_for_testing=y_list_for_testing,
+            x_variable_name=x_variable_name,
+            y_variable_name=y_variable_name,
+            data_width=data_width
+        )
+        expected = [
+            f'report "======Simulation Start======" severity Note;',
+            f"{x_variable_name} <= to_signed({x_list_for_testing[0]},{data_width});",
+            f"wait for 1*clk_period;",
+            f"report \"The value of '{y_variable_name}' is \" & integer'image(to_integer(unsigned({y_variable_name})));",
+            f'assert {y_variable_name}={y_list_for_testing[0]} report "The test case {x_list_for_testing[0]} fail" severity failure;',
+            f"{x_variable_name} <= to_signed({x_list_for_testing[1]},{data_width});",
+            f"wait for 1*clk_period;",
+            f"report \"The value of '{y_variable_name}' is \" & integer'image(to_integer(unsigned({y_variable_name})));",
+            f'assert {y_variable_name}="{y_list_for_testing[1]}" report "The test case {x_list_for_testing[1]} fail" severity failure;',
+            f"{x_variable_name} <= to_signed({x_list_for_testing[2]},{data_width});",
+            f"wait for 1*clk_period;",
+            f"report \"The value of '{y_variable_name}' is \" & integer'image(to_integer(unsigned({y_variable_name})));",
+            f'assert {y_variable_name}={y_list_for_testing[2]} report "The test case {x_list_for_testing[2]} fail" severity failure;',
+            f'report "======Simulation Success======" severity Note;',
+            f'report "Please check the output message." severity Note;',
+            f"wait;"
+        ]
+        actual = list(test_cases_precomputed_scalar_function())
+        self.assertSequenceEqual(expected, actual)
+
+    def test_TestCasesPrecomputedScalarFunction_different_lenghts_of_list(self):
+        x_list_for_testing = [1, 2, 3]
+        y_list_for_testing = [-5, -3]
+        x_variable_name = "x_name"
+        y_variable_name = "y_name"
+        data_width = 3
+        self.assertRaises(
+            AssertionError,
+            TestCasesPrecomputedScalarFunction,
+            x_list_for_testing,
+            y_list_for_testing,
+            x_variable_name,
+            y_variable_name,
+            data_width
+        )
